+++ src/DotRecast.Detour.Crowd/DtCrowdConfig.cs
@@ -23,52 +23,16 @@ namespace DotRecast.Detour.Crowd
     {
         public readonly float maxAgentRadius;
 
-        /**
-     * Max number of path requests in the queue
-     */
-        public int pathQueueSize = 32;
-
-        /**
-     * Max number of sliced path finding iterations executed per update (used to handle longer paths and replans)
-     */
-        public int maxFindPathIterations = 100;
-
-        /**
-     * Max number of sliced path finding iterations executed per agent to find the initial path to target
-     */
-        public int maxTargetFindPathIterations = 20;
-
-        /**
-     * Min time between topology optimizations (in seconds)
-     */
-        public float topologyOptimizationTimeThreshold = 0.5f;
-
-        /**
-     * The number of polygons from the beginning of the corridor to check to ensure path validity
-     */
-        public int checkLookAhead = 10;
-
-        /**
-     * Min time between target re-planning (in seconds)
-     */
-        public float targetReplanDelay = 1.0f;
-
-        /**
-     * Max number of sliced path finding iterations executed per topology optimization per agent
-     */
-        public int maxTopologyOptimizationIterations = 32;
-
+        public int pathQueueSize = 32; // Max number of path requests in the queue
+        public int maxFindPathIterations = 100; // Max number of sliced path finding iterations executed per update (used to handle longer paths and replans)
+        public int maxTargetFindPathIterations = 20; // Max number of sliced path finding iterations executed per agent to find the initial path to target
+        public float topologyOptimizationTimeThreshold = 0.5f; // Min time between topology optimizations (in seconds)
+        public int checkLookAhead = 10; // The number of polygons from the beginning of the corridor to check to ensure path validity
+        public float targetReplanDelay = 1.0f; // Min time between target re-planning (in seconds)
+        public int maxTopologyOptimizationIterations = 32; // Max number of sliced path finding iterations executed per topology optimization per agent
         public float collisionResolveFactor = 0.7f;
-
-        /**
-     * Max number of neighbour agents to consider in obstacle avoidance processing
-     */
-        public int maxObstacleAvoidanceCircles = 6;
-
-        /**
-     * Max number of neighbour segments to consider in obstacle avoidance processing
-     */
-        public int maxObstacleAvoidanceSegments = 8;
+        public int maxObstacleAvoidanceCircles = 6; // Max number of neighbour agents to consider in obstacle avoidance processing
+        public int maxObstacleAvoidanceSegments = 8; // Max number of neighbour segments to consider in obstacle avoidance processing
 
         public DtCrowdConfig(float maxAgentRadius)
         {